# Conflicts:
#	project/build.scala

+++ src/test/scala/mesosphere/marathon/integration/setup/ProcessKeeper.scala
@@ -12,6 +12,7 @@ import org.apache.commons.io.FileUtils
 import org.rogach.scallop.ScallopConf
 import org.slf4j.LoggerFactory
 
+import scala.collection.immutable.ListMap
 import scala.concurrent.duration._
 import scala.concurrent.{ Await, ExecutionContext, Future, Promise }
 import scala.sys.ShutdownHookThread
@@ -27,7 +28,7 @@ import scala.util.{ Failure, Success, Try }
 object ProcessKeeper {
 
   private[this] val log = LoggerFactory.getLogger(getClass.getName)
-  private[this] var processes = List.empty[Process]
+  private[this] var processes = ListMap.empty[String, Process]
   private[this] var services = List.empty[Service]
 
   private[this] val ENV_MESOS_WORK_DIR: String = "MESOS_WORK_DIR"
@@ -35,18 +36,21 @@ object ProcessKeeper {
   def startHttpService(port: Int, assetPath: String) = {
     startService {
       log.info(s"Start Http Service on port $port")
-      val conf = new ScallopConf(Array("--http_port", port.toString, "--assets_path", assetPath)) with HttpConf
-      conf.afterInit()
+      val conf = new ScallopConf(Array("--http_port", port.toString, "--assets_path", assetPath)) with HttpConf {
+        verify()
+      }
       val injector = Guice.createInjector(new MetricsModule, new HttpModule(conf), new HttpServiceTestModule)
       injector.getInstance(classOf[HttpService])
     }
   }
 
-  def startZooKeeper(port: Int, workDir: String) {
-    val args = "org.apache.zookeeper.server.ZooKeeperServerMain" :: port.toString :: workDir :: Nil
+  def startZooKeeper(port: Int, workDir: String, wipeWorkDir: Boolean = true) {
+    val args = "-Dzookeeper.jmx.log4j.disable=true" :: "org.apache.zookeeper.server.ZooKeeperServerMain" :: port.toString :: workDir :: Nil
     val workDirFile = new File(workDir)
-    FileUtils.deleteDirectory(workDirFile)
-    FileUtils.forceMkdir(workDirFile)
+    if (wipeWorkDir) {
+      FileUtils.deleteDirectory(workDirFile)
+      FileUtils.forceMkdir(workDirFile)
+    }
     startJavaProcess("zookeeper", heapInMegs = 256, args, new File("."), sys.env, _.contains("binding to port"))
   }
 
@@ -55,15 +59,50 @@ object ProcessKeeper {
     val mesosWorkDirFile: File = new File(mesosWorkDirForMesos)
     FileUtils.deleteDirectory(mesosWorkDirFile)
     FileUtils.forceMkdir(mesosWorkDirFile)
+
+    val credentialsPath = write(mesosWorkDirFile, fileName = "credentials", content = "principal1 secret1")
+    val aclsPath = write(mesosWorkDirFile, fileName = "acls.json", content =
+      """
+        |{
+        |  "run_tasks": [{
+        |    "principals": { "type": "ANY" },
+        |    "users": { "type": "ANY" }
+        |  }],
+        |  "register_frameworks": [{
+        |    "principals": { "type": "ANY" },
+        |    "roles": { "type": "ANY" }
+        |  }],
+        |  "reserve_resources": [{
+        |    "roles": { "type": "ANY" },
+        |    "principals": { "type": "ANY" },
+        |    "resources": { "type": "ANY" }
+        |  }],
+        |  "create_volumes": [{
+        |    "roles": { "type": "ANY" },
+        |    "principals": { "type": "ANY" },
+        |    "volume_types": { "type": "ANY" }
+        |  }]
+        |}
+      """.stripMargin)
+
+    log.info(s">>> credentialsPath = $credentialsPath")
+    val mesosEnv = Seq(
+      ENV_MESOS_WORK_DIR -> mesosWorkDirForMesos,
+      "MESOS_LAUNCHER" -> "posix",
+      "MESOS_CONTAINERIZERS" -> "docker,mesos",
+      "MESOS_ROLES" -> "public,foo",
+      "MESOS_ACLS" -> s"file://$aclsPath",
+      "MESOS_CREDENTIALS" -> s"file://$credentialsPath")
     startProcess(
       "mesos",
-      Process(Seq("mesos-local", "--ip=127.0.0.1"), cwd = None, ENV_MESOS_WORK_DIR -> mesosWorkDirForMesos),
+      Process(Seq("mesos-local", "--ip=127.0.0.1"), cwd = None, mesosEnv: _*),
       upWhen = _.toLowerCase.contains("registered with master"))
   }
 
   def startMarathon(cwd: File, env: Map[String, String], arguments: List[String],
                     mainClass: String = "mesosphere.marathon.Main",
-                    startupLine: String = "Started ServerConnector"): Process = {
+                    startupLine: String = "Started ServerConnector",
+                    processName: String = "marathon"): Process = {
 
     val debugArgs = List(
       "-Dakka.loglevel=DEBUG",
@@ -71,17 +110,33 @@ object ProcessKeeper {
       "-Dakka.actor.debug.autoreceive=true",
       "-Dakka.actor.debug.lifecycle=true"
     )
-    val argsWithMain = mainClass :: arguments
 
-    val mesosWorkDir: String = "/tmp/marathon-itest-marathon"
-    val mesosWorkDirFile: File = new File(mesosWorkDir)
-    FileUtils.deleteDirectory(mesosWorkDirFile)
-    FileUtils.forceMkdir(mesosWorkDirFile)
+    val marathonWorkDir: String = "/tmp/marathon-itest-marathon"
+    val marathonWorkDirFile: File = new File(marathonWorkDir)
+    FileUtils.deleteDirectory(marathonWorkDirFile)
+    FileUtils.forceMkdir(marathonWorkDirFile)
+
+    val secretPath = write(marathonWorkDirFile, fileName = "marathon-secret", content = "secret1")
+    val authSettings = List(
+      "--mesos_authentication_principal", "principal1",
+      "--mesos_role", "foo",
+      "--mesos_authentication_secret_file", s"$secretPath"
+    )
+
+    val argsWithMain = mainClass :: arguments ++ authSettings
 
     startJavaProcess(
-      "marathon", heapInMegs = 512, debugArgs ++ argsWithMain, cwd,
-      env + (ENV_MESOS_WORK_DIR -> mesosWorkDir),
+      processName, heapInMegs = 512, /* debugArgs ++ */ argsWithMain, cwd,
+      env + (ENV_MESOS_WORK_DIR -> marathonWorkDir),
       upWhen = _.contains(startupLine))
+  }
+
+  private[this] def write(dir: File, fileName: String, content: String): String = {
+    val file = File.createTempFile(fileName, "", dir)
+    file.deleteOnExit()
+    FileUtils.write(file, content)
+    file.setReadable(true)
+    file.getAbsolutePath
   }
 
   def startJavaProcess(name: String, heapInMegs: Int, arguments: List[String],
@@ -89,8 +144,9 @@ object ProcessKeeper {
     val javaExecutable = sys.props.get("java.home").fold("java")(_ + "/bin/java")
     val classPath = sys.props.getOrElse("java.class.path", "target/classes")
     val memSettings = s"-Xmx${heapInMegs}m"
+    // Omit the classpath in order to avoid cluttering the tests output
+    log.info(s"Start java process $name with command: ${(javaExecutable :: memSettings :: arguments).mkString(" ")}")
     val command: List[String] = javaExecutable :: memSettings :: "-classpath" :: classPath :: arguments
-    log.info(s"Start java process $name with command: ${command.mkString(" ")}")
     val builder = Process(command, cwd, env.toList: _*)
     val process = startProcess(name, builder, upWhen)
     log.info(s"Java process $name up and running!")
@@ -98,6 +154,8 @@ object ProcessKeeper {
   }
 
   def startProcess(name: String, processBuilder: ProcessBuilder, upWhen: String => Boolean, timeout: Duration = 30.seconds): Process = {
+    require(!processes.contains(name), s"Process with $name already started")
+
     sealed trait ProcessState
     case object ProcessIsUp extends ProcessState
     case object ProcessExited extends ProcessState
@@ -128,16 +186,17 @@ object ProcessKeeper {
     val upOrExited = Future.firstCompletedOf(Seq(up.future, processExitCode))(ExecutionContext.global)
     Try(Await.result(upOrExited, timeout)) match {
       case Success(result) =>
-        processes = process :: processes
         result match {
           case ProcessExited =>
             throw new IllegalStateException(s"Process $name exited before coming up. Give up. $processBuilder")
-          case ProcessIsUp => log.info(s"Process $name is up and running. ${processes.size} processes in total.")
+          case ProcessIsUp =>
+            processes += name -> process
+            log.info(s"Process $name is up and running. ${processes.size} processes in total.")
         }
       case Failure(_) =>
         process.destroy()
         throw new IllegalStateException(
-          s"Process $name does not came up within time bounds ($timeout). Give up. $processBuilder")
+          s"Process $name did not come up within time bounds ($timeout). Give up. $processBuilder")
     }
     process
   }
@@ -168,38 +227,31 @@ object ProcessKeeper {
     }
   }
 
-  def stopAllProcesses(): Unit = {
-
-    def waitForProcessesToFinish(): Unit = {
-      processes.foreach(p => Try(p.destroy()))
-
+  def stopProcess(name: String): Unit = {
+    import mesosphere.util.ThreadPoolContext.ioContext
+    log.info(s"Stop Process $name")
+    val process = processes(name)
+    def killProcess: Int = {
       // Unfortunately, there seem to be race conditions in Process.exitValue.
       // Thus this ugly workaround.
-      val waitForExitInThread = new Thread() {
-        override def run(): Unit = {
-          processes.foreach(_.exitValue())
+      Await.result(Future {
+        scala.concurrent.blocking {
+          Try(process.destroy())
+          process.exitValue()
         }
-      }
-      waitForExitInThread.start()
-      try {
-        waitForExitInThread.join(1000)
-      }
-      finally {
-        waitForExitInThread.interrupt()
-      }
+      }, 5.seconds)
     }
+    //retry on fail
+    Try(killProcess) recover { case _ => killProcess } match {
+      case Success(value)       => processes -= name
+      case Failure(NonFatal(e)) => log.error("giving up waiting for processes to finish", e)
+    }
+    log.info(s"Stop Process $name: Done")
+  }
 
-    try waitForProcessesToFinish()
-    catch {
-      case NonFatal(e) =>
-        log.error("while waiting for processes to finish", e)
-        try waitForProcessesToFinish()
-        catch {
-          case NonFatal(e) =>
-            log.error("giving up waiting for processes to finish", e)
-        }
-    }
-    processes = Nil
+  def stopAllProcesses(): Unit = {
+    processes.keys.toSeq.reverse.foreach(stopProcess)
+    processes = ListMap.empty
   }
 
   def startService(service: Service): Unit = {
@@ -222,18 +274,15 @@ object ProcessKeeper {
     log.info(s"Cleaning up Processes $processes and Services $services")
     stopAllProcesses()
     stopAllServices()
+    log.info(s"Cleaning up Processes $processes and Services $services: Done")
+  }
+
+  def exitValue(processName: String): Int = {
+    processes(processName).exitValue()
   }
 
   val shutDownHook: ShutdownHookThread = sys.addShutdownHook {
     shutdown()
-  }
-
-  def main(args: Array[String]) {
-    //startMarathon(new File("."), Map("MESOS_NATIVE_LIBRARY" -> "/usr/local/lib/libmesos.dylib"), List("--master", "local", "--event_subscriber", "http_callback"))
-    startZooKeeper(2183, "/tmp/foo")
-    Thread.sleep(10000)
-    stopAllProcesses()
-    //startHttpService(11211, ".")
   }
 }
